fix: remove libero dependencies of non-libero envs
