+++ .github/workflows/embodied-e2e-tests.yml
@@ -23,6 +23,7 @@ jobs:
       - name: ManiSkill PPO test
         timeout-minutes: 20
         run: |
+          unset PYTHONPATH
           export REPO_PATH=$(pwd)
           source .venv/bin/activate
           bash tests/e2e_tests/embodied/run.sh maniskill_ppo_openvla
@@ -51,6 +52,7 @@ jobs:
       - name: ManiSkill GRPO test
         timeout-minutes: 20
         run: |
+          unset PYTHONPATH
           export REPO_PATH=$(pwd)
           source .venv/bin/activate
           cp -r /workspace/dataset/maniskill_assets/assets ${REPO_PATH}/rlinf/envs/maniskill/
@@ -59,6 +61,7 @@ jobs:
       - name: Libero-Goal GRPO test
         timeout-minutes: 20
         run: |
+          unset PYTHONPATH
           export REPO_PATH=$(pwd)
           source .venv/bin/activate
           bash tests/e2e_tests/embodied/run.sh libero_goal_grpo_openvlaoft
@@ -66,6 +69,7 @@ jobs:
       - name: Libero-130 GRPO test
         timeout-minutes: 20
         run: |
+          unset PYTHONPATH
           export REPO_PATH=$(pwd)
           source .venv/bin/activate
           bash tests/e2e_tests/embodied/run.sh libero_130_grpo_openvlaoft
@@ -95,6 +99,7 @@ jobs:
       - name: BEHAVIOR-OpenVLA-OFT test
         timeout-minutes: 30
         run: |
+          unset PYTHONPATH
           export REPO_PATH=$(pwd)
           source .venv/bin/activate
           export OMNIGIBSON_DATA_PATH=/workspace/dataset/behavior-datasets
@@ -126,6 +131,7 @@ jobs:
       - name: Libero-Spatial PPO test
         timeout-minutes: 20
         run: |
+          unset PYTHONPATH
           export REPO_PATH=$(pwd)
           source .venv/bin/activate
           bash tests/e2e_tests/embodied/run.sh libero_spatial_ppo_gr00t
@@ -155,6 +161,7 @@ jobs:
       - name: ISAACLAB PPO test
         timeout-minutes: 20
         run: |
+          unset PYTHONPATH
           export REPO_PATH=$(pwd)
           source .venv/bin/activate
           source /workspace/dataset/isaac-sim-5.1/setup_conda_env.sh
@@ -184,6 +191,7 @@ jobs:
       - name: Libero-Spatial PPO test
         timeout-minutes: 20
         run: |
+          unset PYTHONPATH
           export REPO_PATH=$(pwd)
           source .venv/bin/activate
           bash tests/e2e_tests/embodied/run.sh libero_spatial_ppo_openpi
@@ -191,6 +199,7 @@ jobs:
       - name: Libero-Spatial PPO Pi05 test
         timeout-minutes: 20
         run: |
+          unset PYTHONPATH
           export REPO_PATH=$(pwd)
           source .venv/bin/activate
           bash tests/e2e_tests/embodied/run.sh libero_spatial_ppo_openpi05
@@ -218,6 +227,7 @@ jobs:
       - name: Metaworld PPO test
         timeout-minutes: 20
         run: |
+          unset PYTHONPATH
           export REPO_PATH=$(pwd)
           source .venv/bin/activate
           bash tests/e2e_tests/embodied/run.sh metaworld_50_ppo_openpi
@@ -246,6 +256,7 @@ jobs:
       - name: Calvin PPO test
         timeout-minutes: 20
         run: |
+          unset PYTHONPATH
           export REPO_PATH=$(pwd)
           source .venv/bin/activate
           bash tests/e2e_tests/embodied/run.sh calvin_ppo_openpi
