+++ examples/embodiment/eval_embodiment.sh
@@ -20,9 +20,6 @@ export ISAAC_PATH=${ISAAC_PATH:-/path/to/isaac-sim}
 export EXP_PATH=${EXP_PATH:-$ISAAC_PATH/apps}
 export CARB_APP_PATH=${CARB_APP_PATH:-$ISAAC_PATH/kit}
 
-LIBERO_PATH=/opt/libero
-export PYTHONPATH=${LIBERO_PATH}:$PYTHONPATH
-
 export CUDA_LAUNCH_BLOCKING=1
 export HYDRA_FULL_ERROR=1
