+++ rlinf/envs/libero/utils.py
@@ -15,150 +15,10 @@
 """Utils for evaluating policies in LIBERO simulation environments."""
 
 import math
-import os
-from typing import Optional, Union
+from typing import Union
 
-import imageio
 import libero.libero.benchmark as benchmark
 import numpy as np
-import torch
-from libero.libero.benchmark import Benchmark
-from PIL import Image, ImageDraw, ImageFont
-
-
-def tile_images(
-    images: list[Union[np.ndarray, torch.Tensor]], nrows: int = 1
-) -> Union[np.ndarray, torch.Tensor]:
-    """
-    Copied from maniskill https://github.com/haosulab/ManiSkill
-    Tile multiple images to a single image comprised of nrows and an appropriate number of columns to fit all the images.
-    The images can also be batched (e.g. of shape (B, H, W, C)), but give images must all have the same batch size.
-
-    if nrows is 1, images can be of different sizes. If nrows > 1, they must all be the same size.
-    """
-    # Sort images in descending order of vertical height
-    batched = False
-    if len(images[0].shape) == 4:
-        batched = True
-    if nrows == 1:
-        images = sorted(images, key=lambda x: x.shape[0 + batched], reverse=True)
-
-    columns: list[list[Union[np.ndarray, torch.Tensor]]] = []
-    if batched:
-        max_h = images[0].shape[1] * nrows
-        cur_h = 0
-        cur_w = images[0].shape[2]
-    else:
-        max_h = images[0].shape[0] * nrows
-        cur_h = 0
-        cur_w = images[0].shape[1]
-
-    # Arrange images in columns from left to right
-    column = []
-    for im in images:
-        if cur_h + im.shape[0 + batched] <= max_h and cur_w == im.shape[1 + batched]:
-            column.append(im)
-            cur_h += im.shape[0 + batched]
-        else:
-            columns.append(column)
-            column = [im]
-            cur_h, cur_w = im.shape[0 + batched : 2 + batched]
-    columns.append(column)
-
-    # Tile columns
-    total_width = sum(x[0].shape[1 + batched] for x in columns)
-
-    is_torch = False
-    if torch is not None:
-        is_torch = isinstance(images[0], torch.Tensor)
-
-    output_shape = (max_h, total_width, 3)
-    if batched:
-        output_shape = (images[0].shape[0], max_h, total_width, 3)
-    if is_torch:
-        output_image = torch.zeros(output_shape, dtype=images[0].dtype)
-    else:
-        output_image = np.zeros(output_shape, dtype=images[0].dtype)
-    cur_x = 0
-    for column in columns:
-        cur_w = column[0].shape[1 + batched]
-        next_x = cur_x + cur_w
-        if is_torch:
-            column_image = torch.concatenate(column, dim=0 + batched)
-        else:
-            column_image = np.concatenate(column, axis=0 + batched)
-        cur_h = column_image.shape[0 + batched]
-        output_image[..., :cur_h, cur_x:next_x, :] = column_image
-        cur_x = next_x
-    return output_image
-
-
-def put_text_on_image(
-    image: np.ndarray, lines: list[str], max_width: int = 200
-) -> np.ndarray:
-    """
-    Put text lines on an image with automatic line wrapping.
-
-    Args:
-        image: Input image as numpy array
-        lines: List of text lines to add
-        max_width: Maximum width for text wrapping
-    """
-    assert image.dtype == np.uint8, image.dtype
-    image = image.copy()
-    image = Image.fromarray(image)
-    draw = ImageDraw.Draw(image)
-    font = ImageFont.load_default(size=20)
-
-    new_lines = []
-    for line in lines:
-        words = line.split()
-        current_line = []
-
-        for word in words:
-            test_line = " ".join(current_line + [word])
-            test_width = font.getlength(test_line)
-
-            if test_width <= max_width:
-                current_line.append(word)
-            else:
-                new_lines.append(" ".join(current_line))
-                current_line = [word]
-        if current_line:
-            new_lines.append(" ".join(current_line))
-
-    y = -10
-    for line in new_lines:
-        bbox = draw.textbbox((0, 0), text=line)
-        textheight = bbox[3] - bbox[1]
-        y += textheight + 10
-        x = 10
-        draw.text((x, y), text=line, fill=(0, 0, 0))
-    return np.array(image)
-
-
-def put_info_on_image(
-    image: np.ndarray,
-    info: dict[str, float],
-    extras: Optional[list[str]] = None,
-    overlay: bool = True,
-) -> np.ndarray:
-    """
-    Put information dictionary and extra lines on an image.
-
-    Args:
-        image: Input image
-        info: Dictionary of key-value pairs to display
-        extras: Additional text lines to display
-        overlay: Whether to overlay text on image
-    """
-    lines = [
-        f"{k}: {v:.3f}" if isinstance(v, float) else f"{k}: {v}"
-        for k, v in info.items()
-    ]
-    if extras is not None:
-        lines.extend(extras)
-    return put_text_on_image(image, lines)
 
 
 def get_libero_image(obs: dict[str, np.ndarray]) -> np.ndarray:
@@ -221,27 +81,7 @@ def quat2axisangle(quat: np.ndarray) -> np.ndarray:
     return (quat[:3] * 2.0 * math.acos(quat[3])) / den
 
 
-def save_rollout_video(
-    rollout_images: list[np.ndarray], output_dir: str, video_name: str, fps: int = 30
-) -> None:
-    """
-    Saves an MP4 replay of an episode.
-
-    Args:
-        rollout_images: List of images from the episode
-        output_dir: Directory to save the video
-        video_name: Name of the output video file
-        fps: Frames per second for the video
-    """
-    os.makedirs(output_dir, exist_ok=True)
-    mp4_path = os.path.join(output_dir, f"{video_name}.mp4")
-    video_writer = imageio.get_writer(mp4_path, fps=fps)
-    for img in rollout_images:
-        video_writer.append_data(img)
-    video_writer.close()
-
-
-def get_benchmark_overridden(benchmark_name) -> Benchmark:
+def get_benchmark_overridden(benchmark_name) -> benchmark.Benchmark:
     """
     Return the Benchmark class for a given name.
     For "libero_130": return a dynamically aggregated class from all suites.
@@ -269,7 +109,7 @@ def get_benchmark_overridden(benchmark_name) -> Benchmark:
             if task_name not in aggregated_task_map:
                 aggregated_task_map[task_name] = task
 
-    class LIBERO_ALL(Benchmark):
+    class LIBERO_ALL(benchmark.Benchmark):
         def __init__(self, task_order_index=0):
             super().__init__(task_order_index=task_order_index)
             self.name = "libero_130"
