+++ rlinf/envs/metaworld/venv.py
@@ -19,7 +19,8 @@ from typing import Any, Callable, Optional, Union
 
 import gymnasium as gym
 import numpy as np
-from libero.libero.envs.venv import (
+
+from rlinf.envs.venv import (
     BaseVectorEnv,
     CloudpickleWrapper,
     EnvWorker,
